Update to reduce the chances of bucket overflow error in  (#56)
Updated the code computing the value of hash table size, to reduce the chances of bucket overflow error happening in hashmap.cu. When n is small, the effect of uneven distribution of hash values is more pronounced, and we select the table size to be 4 * nextPow2 and for larger n, we take it to be 2 * nextPow2. This way, for larger n, when memory size consideration is more important, the table_size gets computed to be the same as it used to be.

Co-authored-by: Venkat Ajjanagadde <veajjana@microsoft.com>